refactor(core): move computer use contracts to tool contracts
