+++ src/crates/assembly/core/src/util/errors.rs
@@ -179,6 +179,12 @@ impl From<bitfun_agent_stream::StreamProcessorError> for BitFunError {
     }
 }
 
+impl From<bitfun_agent_tools::computer_use::ComputerUseContractError> for BitFunError {
+    fn from(error: bitfun_agent_tools::computer_use::ComputerUseContractError) -> Self {
+        Self::Tool(error.to_string())
+    }
+}
+
 #[cfg(feature = "service-integrations")]
 impl From<bitfun_services_integrations::mcp::MCPRuntimeError> for BitFunError {
     fn from(error: bitfun_services_integrations::mcp::MCPRuntimeError) -> Self {
